requirements.txt
streamlit
assemblyai
python-dotenv
google-generativeai
